requirements.txt
streamlit
python-pptx
easyocr
opencv-python-headless
numpy
Pillow
